Applying the naming of CEWorkspaceFile + Manager to every file
