--- CodeEdit/Features/FileSystemClient/Models/FileSystemItem.swift
@@ -1,480 +1,0 @@
-//
-//  FileItem.swift
-//  CodeEdit
-//
-//  Created by Matthijs Eikelenboom on 07/02/2023.
-//
-
-import Foundation
-import SwiftUI
-import UniformTypeIdentifiers
-
-extension FileSystemClient {
-
-    enum FileItemCodingKeys: String, CodingKey {
-        case id
-        case url
-        case children
-        case changeType
-    }
-
-    /// An object containing all necessary information and actions for a specific file in the workspace
-    final class FileSystemItem: Codable, Comparable, Hashable, Identifiable, TabBarItemRepresentable {
-
-        /// The id of the ``FileSystemClient/FileSystemClient/FileItem``.
-        ///
-        /// This is equal to `url.relativePath`
-        var id: String { url.relativePath }
-
-        /// Returns the file name (e.g.: `Package.swift`)
-        var name: String { url.lastPathComponent }
-
-        /// Returns the extension of the file or an empty string if no extension is present.
-        var type: FileIcon.FileType { .init(rawValue: url.pathExtension) ?? .txt }
-
-        /// Returns the URL of the ``FileSystemClient/FileSystemClient/FileItem``
-        var url: URL
-
-        /// Return the icon of the file as `Image`
-        var icon: Image { Image(systemName: systemImage) }
-
-        /// Returns the children of the current ``FileSystemClient/FileSystemClient/FileItem``.
-        ///
-        /// If the current ``FileSystemClient/FileSystemClient/FileItem`` is a file this will be `nil`.
-        /// If it is an empty folder this will be an empty array.
-        var children: [FileSystemItem]?
-
-        /// Returns a parent ``FileSystemClient/FileSystemClient/FileItem``.
-        ///
-        /// If the item already is the top-level ``FileSystemClient/FileSystemClient/FileItem`` this returns `nil`.
-        var parent: FileSystemItem?
-
-        /// Returns the `id` in ``TabBarItemID`` enum form
-        var tabID: TabBarItemID { .codeEditor(id) }
-
-        var fileIdentifier = UUID().uuidString
-
-        var watcher: DispatchSourceFileSystemObject?
-        var watcherCode: ((FileSystemItem) -> Void)?
-
-        /// Returns the Git status of a file as ``GitType``
-        var gitStatus: GitType?
-
-        /// Returns a boolean that is true if ``children`` is not `nil`
-        var isFolder: Bool { url.hasDirectoryPath }
-
-        /// Returns a boolean that is true if the file item is the root folder of the workspace.
-        var isRoot: Bool { parent == nil }
-
-        /// Returns a boolean that is true if the file item actually exists in the file system
-        var doesExist: Bool { FileSystemItem.fileManger.fileExists(atPath: self.url.path) }
-
-        /// Returns a string describing a SFSymbol for the current ``FileSystemClient/FileSystemClient/FileItem``
-        ///
-        /// Use it like this
-        /// ```swift
-        /// Image(systemName: item.systemImage)
-        /// ```
-        var systemImage: String {
-            if let children = children {
-                // item is a folder
-                return folderIcon(children)
-            } else {
-                // item is a file
-                return FileIcon.fileIcon(fileType: type)
-            }
-        }
-
-        var debugFileHeirachy: String { childrenDescription(tabCount: 0) }
-
-        init(
-            url: URL,
-            children: [FileSystemItem]? = nil,
-            changeType: GitType? = nil
-        ) {
-            self.url = url
-            self.children = children
-            self.gitStatus = changeType
-        }
-
-        required init(from decoder: Decoder) throws {
-            let values = try decoder.container(keyedBy: FileItemCodingKeys.self)
-            url = try values.decode(URL.self, forKey: .url)
-            children = try values.decode([FileSystemItem]?.self, forKey: .children)
-            gitStatus = try values.decode(GitType.self, forKey: .changeType)
-        }
-
-        func encode(to encoder: Encoder) throws {
-            var container = encoder.container(keyedBy: FileItemCodingKeys.self)
-            try container.encode(id, forKey: .id)
-            try container.encode(url, forKey: .url)
-            try container.encode(children, forKey: .children)
-            try container.encode(gitStatus, forKey: .changeType)
-        }
-
-        func activateWatcher() -> Bool {
-            // check that there is watcher code and that opening the file succeeded
-            guard let watcherCode else { return false }
-            let descriptor = open(self.url.path, O_EVTONLY)
-            guard descriptor > 0 else { return false }
-
-            // create the source
-            let source = DispatchSource.makeFileSystemObjectSource(
-                fileDescriptor: descriptor,
-                eventMask: .write,
-                queue: DispatchQueue.global()
-            )
-            if descriptor > 2000 {
-                print("Watcher \(descriptor) used up on \(url.path)")
-            }
-            source.setEventHandler { watcherCode(self) }
-            source.setCancelHandler { close(descriptor) }
-            source.resume()
-            self.watcher = source
-
-            // TODO: reindex the current item, because the files in the item may have changed
-            // since the initial load on startup.
-            return true
-        }
-
-        func childrenDescription(tabCount: Int) -> String {
-            var myDetails = "\(String(repeating: "|  ", count: max(tabCount - 1, 0)))\(tabCount != 0 ? "╰--" : "")"
-            myDetails += "\(url.path)"
-            if !self.isFolder { // if im a file, just return the url
-                return myDetails
-            } else { // if im a folder, return the url and its children's details
-                var childDetails = "\(myDetails)"
-                for child in children ?? [] {
-                    childDetails += "\n\(child.childrenDescription(tabCount: tabCount + 1))"
-                }
-                return childDetails
-            }
-        }
-
-        /// Returns a string describing a SFSymbol for folders
-        ///
-        /// If it is the top-level folder this will return `"square.dashed.inset.filled"`.
-        /// If it is a `.codeedit` folder this will return `"folder.fill.badge.gearshape"`.
-        /// If it has children this will return `"folder.fill"` otherwise `"folder"`.
-        private func folderIcon(_ children: [FileSystemItem]) -> String {
-            if self.parent == nil {
-                return "square.dashed.inset.filled"
-            }
-            if self.name == ".codeedit" {
-                return "folder.fill.badge.gearshape"
-            }
-            return children.isEmpty ? "folder" : "folder.fill"
-        }
-
-        /// Returns the file name with optional extension (e.g.: `Package.swift`)
-        func fileName(typeHidden: Bool) -> String {
-            typeHidden ? url.deletingPathExtension().lastPathComponent : name
-        }
-
-        /// Return the file's UTType
-        var contentType: UTType? {
-            try? url.resourceValues(forKeys: [.contentTypeKey]).contentType
-        }
-
-        /// Returns a `Color` for a specific `fileType`
-        ///
-        /// If not specified otherwise this will return `Color.accentColor`
-        var iconColor: Color {
-            FileIcon.iconColor(fileType: type)
-        }
-
-        // MARK: Statics
-        /// The default `FileManager` instance
-        static let fileManger = FileManager.default
-
-        // MARK: Intents
-        /// Allows the user to view the file or folder in the finder application
-        func showInFinder() {
-            NSWorkspace.shared.activateFileViewerSelecting([url])
-        }
-
-        /// Allows the user to launch the file or folder as it would be in finder
-        func openWithExternalEditor() {
-            NSWorkspace.shared.open(url)
-        }
-        
-        /// Flattens the children of `FileSystemItem` recursively with depth.
-        /// - Parameters:
-        ///   - depth: An int
-        ///   - ignoringFolders: <#ignoringFolders description#>
-        /// - Returns: <#description#>
-        func flattenedChildren(withDepth depth: Int, ignoringFolders: Bool) -> [FileSystemItem] {
-            guard depth > 0 else { return [] }
-            guard self.isFolder else { return [self] }
-            var childItems: [FileSystemItem] = ignoringFolders ? [] : [self]
-            children?.forEach { child in
-                childItems.append(contentsOf: child.flattenedChildren(
-                    withDepth: depth - 1,
-                    ignoringFolders: ignoringFolders
-                ))
-            }
-            return childItems
-        }
-
-        
-        /// Returns a list of `FileSystemItems` that are sibilings of ``self``. The `height` parameter lets te function navigate up the
-        /// folder hierarchy to select a starting point from which it should start flettening the items.
-        /// - Parameters:
-        ///   - height: `Int` that tells where to start in the hierarchy
-        ///   - ignoringFolders: Wether the sibling folders should be flattened
-        /// - Returns: A list of `FileSystemItems`
-        func flattenedSiblings(withHeight height: Int, ignoringFolders: Bool) -> [FileSystemItem] {
-            let topMostParent = self.getParent(withHeight: height)
-            return topMostParent.flattenedChildren(withDepth: height, ignoringFolders: ignoringFolders)
-        }
-
-        /// Using the current instance of `FileSystemItem` it will walk back up the Workspace file hiarchy
-        /// the amount of times specified with the `withHeight` parameter.
-        /// - Parameter height: The amount of times you want to up a folder.
-        /// - Returns: The found `FileSystemItem` object, This should always be a folder.
-        private func getParent(withHeight height: Int) -> FileSystemItem {
-            var topmostParent = self
-            for _ in 0..<height {
-                guard let parent = topmostParent.parent else { break }
-                topmostParent = parent
-            }
-
-            return topmostParent
-        }
-
-        /// Recursive function that returns the number of children
-        /// that contain the `searchString` in their path or their subitems' paths.
-        /// Returns `0` if the item is not a folder.
-        /// - Parameters:
-        ///   - searchString: The string
-        ///   - ignoredStrings: The prefixes to ignore if they prefix file names
-        /// - Returns: The number of children that match the conditiions
-        func appearanceWithinChildrenOf(searchString: String, ignoredStrings: [String] = [".", "~"]) -> Int {
-            var count = 0
-            guard self.isFolder else { return 0 }
-            for child in self.children ?? [] {
-                var isIgnored: Bool = false
-                for ignoredString in ignoredStrings where child.name.hasPrefix(ignoredString) {
-                    isIgnored = true // can use regex later
-                }
-
-                if isIgnored {
-                    continue
-                }
-
-                guard !searchString.isEmpty else { count += 1; continue }
-                if child.isFolder {
-                    count += child.appearanceWithinChildrenOf(searchString: searchString) > 0 ? 1 : 0
-                } else {
-                    count += child.name.lowercased().contains(searchString.lowercased()) ? 1 : 0
-                }
-            }
-            return count
-        }
-
-        /// Function that returns an array of the children
-        /// that contain the `searchString` in their path or their subitems' paths.
-        /// Similar to `appearanceWithinChildrenOf(searchString: String)`
-        /// Returns `[]` if the item is not a folder.
-        /// - Parameter searchString: The string
-        /// - Parameter ignoredStrings: The prefixes to ignore if they prefix file names
-        /// - Returns: The children that match the conditiions
-        func childrenSatisfying(searchString: String, ignoredStrings: [String] = [".", "~"]) -> [FileSystemItem] {
-            var satisfyingChildren: [FileSystemItem] = []
-            guard self.isFolder else { return [] }
-            for child in self.children ?? [] {
-                var isIgnored: Bool = false
-                for ignoredString in ignoredStrings where child.name.hasPrefix(ignoredString) {
-                    isIgnored = true // can use regex later
-                }
-
-                if isIgnored {
-                    continue
-                }
-
-                guard !searchString.isEmpty else { satisfyingChildren.append(child); continue }
-                if child.isFolder {
-                    if child.appearanceWithinChildrenOf(searchString: searchString) > 0 {
-                        satisfyingChildren.append(child)
-                    }
-                } else {
-                    if child.name.lowercased().contains(searchString.lowercased()) {
-                        satisfyingChildren.append(child)
-                    }
-                }
-            }
-            return satisfyingChildren
-        }
-
-        /// This function allows creation of folders in the main directory or sub-folders
-        /// - Parameter folderName: The name of the new folder
-        func addFolder(folderName: String) {
-            // Check if folder, if it is create folder under self, else create on same level.
-            var folderUrl = (self.isFolder ?
-                             self.url.appendingPathComponent(folderName) :
-                                self.url.deletingLastPathComponent().appendingPathComponent(folderName))
-
-            // If a file/folder with the same name exists, add a number to the end.
-            var fileNumber = 0
-            while FileSystemItem.fileManger.fileExists(atPath: folderUrl.path) {
-                fileNumber += 1
-                folderUrl = folderUrl.deletingLastPathComponent().appendingPathComponent("\(folderName)\(fileNumber)")
-            }
-
-            // Create the folder
-            do {
-                try FileSystemItem.fileManger.createDirectory(
-                    at: folderUrl,
-                    withIntermediateDirectories: true,
-                    attributes: [:]
-                )
-            } catch {
-                fatalError(error.localizedDescription)
-            }
-        }
-
-        /// This function allows creating files in the selected folder or project main directory
-        /// - Parameter fileName: The name of the new file
-        func addFile(fileName: String) {
-            // check the folder for other files, and see what the most common file extension is
-            var fileExtensions: [String: Int] = ["": 0]
-
-            for child in (self.isFolder ?
-                          self.flattenedSiblings(withHeight: 2, ignoringFolders: true) :
-                          parent?.flattenedSiblings(withHeight: 2, ignoringFolders: true)) ?? [] where !child.isFolder {
-                // if the file extension was present before, add it now
-                let childFileName = child.fileName(typeHidden: false)
-                if let index = childFileName.lastIndex(of: ".") {
-                    let childFileExtension = ".\(childFileName.suffix(from: index).dropFirst())"
-                    fileExtensions[childFileExtension] = (fileExtensions[childFileExtension] ?? 0) + 1
-                } else {
-                    fileExtensions[""] = (fileExtensions[""] ?? 0) + 1
-                }
-            }
-
-            var largestValue = 0
-            var idealExtension = ""
-            for (extName, count) in fileExtensions where count > largestValue {
-                idealExtension = extName
-                largestValue = count
-            }
-
-            var fileUrl = nearestFolder.appendingPathComponent("\(fileName)\(idealExtension)")
-            // If a file/folder with the same name exists, add a number to the end.
-            var fileNumber = 0
-            while FileSystemItem.fileManger.fileExists(atPath: fileUrl.path) {
-                fileNumber += 1
-                fileUrl = fileUrl.deletingLastPathComponent()
-                    .appendingPathComponent("\(fileName)\(fileNumber)\(idealExtension)")
-            }
-
-            // Create the file
-            FileSystemItem.fileManger.createFile(
-                atPath: fileUrl.path,
-                contents: nil,
-                attributes: [FileAttributeKey.creationDate: Date()]
-            )
-        }
-
-        /// Nearest folder refers to the parent directory if this is a non-folder item, or itself if the item is a folder.
-        var nearestFolder: URL {
-            (self.isFolder ?
-                        self.url :
-                        self.url.deletingLastPathComponent())
-        }
-
-        /// This function deletes the item or folder from the current project
-        func delete() {
-            // This function also has to account for how the
-            // - file system can change outside of the editor
-            let deleteConfirmation = NSAlert()
-            let message = "\(String(describing: self.fileName))\(self.isFolder ? " and its children" : "")"
-            deleteConfirmation.messageText = "Do you want to move \(message) to the bin?"
-            deleteConfirmation.alertStyle = .critical
-            deleteConfirmation.addButton(withTitle: "Delete")
-            deleteConfirmation.buttons.last?.hasDestructiveAction = true
-            deleteConfirmation.addButton(withTitle: "Cancel")
-            if deleteConfirmation.runModal() == .alertFirstButtonReturn { // "Delete" button
-                if FileSystemItem.fileManger.fileExists(atPath: self.url.path) {
-                    do {
-                        try FileSystemItem.fileManger.removeItem(at: self.url)
-                    } catch {
-                        fatalError(error.localizedDescription)
-                    }
-                }
-            }
-        }
-
-        /// This function duplicates the item or folder
-        public func duplicate() {
-            // If a file/folder with the same name exists, add "copy" to the end
-            var fileUrl = self.url
-            while FileSystemItem.fileManger.fileExists(atPath: fileUrl.path) {
-                let previousName = fileUrl.lastPathComponent
-                let fileExtension = fileUrl.pathExtension.isEmpty ? "" : ".\(fileUrl.pathExtension)"
-                let fileName = fileExtension.isEmpty ? previousName :
-                    previousName.replacingOccurrences(of: ".\(fileExtension)", with: "")
-                fileUrl = fileUrl.deletingLastPathComponent().appendingPathComponent("\(fileName) copy\(fileExtension)")
-            }
-//            Log.info("Duplicating file to \(fileUrl)")
-
-            if FileSystemItem.fileManger.fileExists(atPath: self.url.path) {
-                do {
-                    try FileSystemItem.fileManger.copyItem(at: self.url, to: fileUrl)
-                } catch {
-//                    Log.error("Error at \(self.url.path) to \(fileUrl.path)")
-                    fatalError(error.localizedDescription)
-                }
-            }
-        }
-
-        /// This function moves the item or folder if possible
-        func move(to newLocation: URL) {
-            guard !FileSystemItem.fileManger.fileExists(atPath: newLocation.path) else { return }
-            createMissingParentDirectory(for: newLocation.deletingLastPathComponent())
-
-            do {
-//                Log.info("Moving file \(self.url.debugDescription) to \(newLocation.debugDescription)")
-                try FileSystemItem.fileManger.moveItem(at: self.url, to: newLocation)
-            } catch { fatalError(error.localizedDescription) }
-
-            // This function recursively creates missing directories if the file is moved to a directory that does not exist
-            func createMissingParentDirectory(for url: URL, createSelf: Bool = true) {
-                // if the folder's parent folder doesn't exist, create it.
-                if !FileSystemItem.fileManger.fileExists(atPath: url.deletingLastPathComponent().path) {
-                    createMissingParentDirectory(for: url.deletingLastPathComponent())
-                }
-                // if the folder doesn't exist and the function was ordered to create it, create it.
-                if createSelf && !FileSystemItem.fileManger.fileExists(atPath: url.path) {
-//                    Log.info("Creating folder \(url.debugDescription)")
-                    // Create the folder
-                    do {
-                        try FileSystemItem.fileManger.createDirectory(at: url,
-                                                                withIntermediateDirectories: true,
-                                                                attributes: [:])
-                    } catch {
-                        fatalError(error.localizedDescription)
-                    }
-                }
-            }
-        }
-
-        // MARK: Comparable
-
-        static func == (lhs: FileSystemClient.FileSystemItem, rhs: FileSystemClient.FileSystemItem) -> Bool {
-            lhs.id == rhs.id
-        }
-
-        static func < (lhs: FileSystemClient.FileSystemItem, rhs: FileSystemClient.FileSystemItem) -> Bool {
-            lhs.url.lastPathComponent < rhs.url.lastPathComponent
-        }
-
-        // MARK: Hashable
-
-        func hash(into hasher: inout Hasher) {
-            hasher.combine(fileIdentifier)
-            hasher.combine(id)
-        }
-
-    }
-}
