+++ CodeEdit/Features/NavigatorSidebar/ProjectNavigator/OutlineView/OutlineView.swift
@@ -19,7 +19,7 @@ struct OutlineView: NSViewControllerRepresentable {
 
     // This is mainly just used to trigger a view update.
     @Binding
-    var selection: WorkspaceClient.FileItem?
+    var selection: CEWorkspaceFile?
 
     typealias NSViewControllerType = OutlineViewController
 
@@ -54,7 +54,7 @@ struct OutlineView: NSViewControllerRepresentable {
 
             listener = workspace.listenerModel.$highlightedFileItem
                 .sink(receiveValue: { [weak self] fileItem in
-                guard let fileItem = fileItem else {
+                guard let fileItem else {
                     return
                 }
                 self?.controller?.reveal(fileItem)
