+++ CodeEdit/Features/NavigatorSidebar/ProjectNavigator/OutlineView/OutlintViewController+OutlineTableViewCellDelegate.swift
@@ -10,7 +10,7 @@ import Foundation
 // MARK: - OutlineTableViewCellDelegate
 
 extension OutlineViewController: OutlineTableViewCellDelegate {
-    func moveFile(file: Item, to destination: URL) {
+    func moveFile(file: CEWorkspaceFile, to destination: URL) {
         if !file.isFolder {
             workspace?.tabManager.tabGroups.closeAllTabs(of: file)
         }
@@ -20,7 +20,7 @@ extension OutlineViewController: OutlineTableViewCellDelegate {
         }
     }
 
-    func copyFile(file: WorkspaceClient.FileItem, to destination: URL) {
-        file.duplicate(to: destination)
+    func copyFile(file: CEWorkspaceFile, to destination: URL) {
+        file.duplicate()
     }
 }
